convert to interface + implementation
