+++ src/PhotoLabeler.PhotoStorageReader/Interfaces/IPhotoReader.cs
@@ -1,0 +1,8 @@
+
+namespace PhotoLabeler.PhotoStorageReader.Interfaces
+{
+	public interface IPhotoReader
+	{
+		string GetImgSrc(string path);
+	}
+}
