+++ src/PhotoLabeler.ServiceLibrary/PhotoLabeler.ServiceLibrary.csproj
@@ -6,12 +6,12 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Localization.Abstractions" Version="3.1.5" />
-    <PackageReference Include="SixLabors.ImageSharp" Version="1.0.0-rc0003" />
     <PackageReference Include="XmpCore" Version="6.1.10" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\PhotoLabeler.Entities\PhotoLabeler.Entities.csproj" />
+    <ProjectReference Include="..\PhotoLabeler.PhotoStorageReader\PhotoLabeler.PhtotoStorageReader.csproj" />
   </ItemGroup>
 
   <ItemGroup>
