+++ src/PhotoLabeler.sln
@@ -23,6 +23,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Lib", "Lib", "{1DDB4502-27C
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoLabeler.Data", "PhotoLabeler.Data\PhotoLabeler.Data.csproj", "{45F73CB1-8383-4A53-AFE8-1E421A512CD0}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoLabeler.PhtotoStorageReader", "PhotoLabeler.PhotoStorageReader\PhotoLabeler.PhtotoStorageReader.csproj", "{65AA5036-A191-49BA-83F4-B45B6CF1217B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -53,6 +55,10 @@ Global
 		{45F73CB1-8383-4A53-AFE8-1E421A512CD0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{45F73CB1-8383-4A53-AFE8-1E421A512CD0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{45F73CB1-8383-4A53-AFE8-1E421A512CD0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{65AA5036-A191-49BA-83F4-B45B6CF1217B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{65AA5036-A191-49BA-83F4-B45B6CF1217B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{65AA5036-A191-49BA-83F4-B45B6CF1217B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{65AA5036-A191-49BA-83F4-B45B6CF1217B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
