+++ src/PhotoLabeler/PhotoLabeler.csproj
@@ -15,6 +15,7 @@
     <ProjectReference Include="..\PhotoLabeler.Data\PhotoLabeler.Data.csproj" />
     <ProjectReference Include="..\PhotoLabeler.Entities\PhotoLabeler.Entities.csproj" />
     <ProjectReference Include="..\PhotoLabeler.ServiceLibrary\PhotoLabeler.ServiceLibrary.csproj" />
+    <ProjectReference Include="..\PhotoLabeler.PhotoStorageReader\PhotoLabeler.PhtotoStorageReader.csproj" />
   </ItemGroup>
   <ItemGroup>
     <Content Update="electron.manifest.json">
